cdh3u4

+++ src/hdfs/org/apache/hadoop/hdfs/server/common/Storage.java
@@ -35,9 +35,6 @@ import org.apache.hadoop.hdfs.protocol.FSConstants;
 import org.apache.hadoop.hdfs.server.common.HdfsConstants.NodeType;
 import org.apache.hadoop.hdfs.server.common.HdfsConstants.StartupOption;
 import org.apache.hadoop.fs.FileUtil;
-import org.apache.hadoop.util.VersionInfo;
-
-
 
 /**
  * Storage information file.
@@ -608,8 +605,12 @@ public abstract class Storage extends StorageInfo {
      * @throws IOException if locking fails.
      */
     FileLock tryLock() throws IOException {
+      boolean deletionHookAdded = false;
       File lockF = new File(root, STORAGE_FILE_LOCK);
-      lockF.deleteOnExit();
+      if (!lockF.exists()) {
+        lockF.deleteOnExit();
+        deletionHookAdded = true;
+      }
       RandomAccessFile file = new RandomAccessFile(lockF, "rws");
       FileLock res = null;
       try {
@@ -621,6 +622,12 @@ public abstract class Storage extends StorageInfo {
         LOG.error("Cannot create lock on " + lockF, e);
         file.close();
         throw e;
+      }
+      if (res != null && !deletionHookAdded) {
+        // If the file existed prior to our startup, we didn't
+        // call deleteOnExit above. But since we successfully locked
+        // the dir, we can take care of cleaning it up.
+        lockF.deleteOnExit();
       }
       return res;
     }
@@ -822,10 +829,6 @@ public abstract class Storage extends StorageInfo {
       }
     }
     return false;
-  }
-
-  public static String getBuildVersion() {
-    return VersionInfo.getRevision();
   }
 
   public static String getRegistrationID(StorageInfo storage) {
